Rename abstractPercussionSet files.
