+++ SeqWiresLib/CMakeLists.txt
@@ -34,7 +34,7 @@ SET( SEQWIRESLIB_SRCS
 	Types/Track/trackTypeConstructor.cpp
 	chord.cpp
 	musicTypes.cpp
-	Percussion/abstractPercussionSet.cpp
+	Percussion/percussionTypeTag.cpp
 	Percussion/builtInPercussionInstruments.cpp
 	Percussion/percussionSetWithPitchMap.cpp
 	pitch.cpp
