+++ Tests/SeqWiresLib/percussionMapProcessorTest.cpp
@@ -1,7 +1,6 @@
 #include <gtest/gtest.h>
 
 #include <SeqWiresLib/Functions/percussionMapFunction.hpp>
-#include <SeqWiresLib/Percussion/abstractPercussionSet.hpp>
 #include <SeqWiresLib/Percussion/builtInPercussionInstruments.hpp>
 #include <SeqWiresLib/Processors/percussionMapProcessor.hpp>
 #include <SeqWiresLib/Types/Track/TrackEvents/percussionEvents.hpp>
